rename intersection -> intersect
