+++ code/src/stf/include/interface/stf/alg/clipping.hpp
@@ -1,0 +1,15 @@
+#pragma once
+
+#include "stf/geom/polyline.hpp"
+
+/**
+ * @file clipping.hpp
+ * @brief A file containing functions that clip geometric objects
+ */
+
+namespace stf::alg
+{
+
+
+
+} // stf::alg
