+++ code/src/stf/include/interface/stf/alg/intersect.hpp
